gmail_assistant/command_processor.py
"""Process voice commands and map to actions."""
import nltk
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords

class CommandProcessor:
    def __init__(self):
        self.commands = {
            'read': self._handle_read,
            'send': self._handle_send,
            'delete': self._handle_delete,
            'search': self._handle_search
        }
        
    def process_command(self, command_text):
        tokens = word_tokenize(command_text.lower())
        command_words = [word for word in tokens 
                        if word not in stopwords.words('english')]
        
        for cmd in self.commands:
            if cmd in command_words:
                return self.commands[cmd](command_words)
        
        return "Command not recognized"

    def _handle_read(self, words):
        # Implement read logic
        pass

    def _handle_send(self, words):
        # Implement send logic
        pass

    def _handle_delete(self, words):
        # Implement delete logic
        pass

    def _handle_search(self, words):
        # Implement search logic
        pass 